Update session_id and patient_id for MiExpressCare
patient_id was pointing to the primary key value for the patient record
in Eko database. But this did not help MiExpressCare to associate the
returned metadata with their sessions in their database. The session_id
was updated to point to the patient.identifier (which is being set by
MiExpressCare to an id that on their end is associated with the session,
likely the primary key) while the patient_id was added and points to our
record's primary key value now for consistency with our documentation